requirements.txt
git+https://github.com/facebookresearch/detectron2.git
pytorch-lightning==2.0.2
smplx==0.1.28
pyrender==0.1.45
opencv-python==4.10.0.84
yacs==0.1.8
scikit-image==0.16.2
einops==0.8.0
timm==1.0.12
pandas==0.25.3
loguru==0.7.3
numpy==1.24.4
chumpy
pyrootutils==1.0.4
timm==1.0.12
einops==0.8.0
